Move shp conversion files, create in3Utils module directory
     adjust code accordingly
     clean code in com2AB some
     clean code in utils, change prints to logs
